GenAI_SourceCode/09consume_llm/requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-anthropic
langchain-groq
python-dotenv
streamlit